feat(agent): improve sub-agent task runtime
